requirements.txt
jinja2~=3.1.6
ruamel.yaml~=0.18.10
werkzeug~=3.1.3
